Add in API Scaffolding
Rename mappings to organizations, as that's what they are.  Give
providers a type, so we can drive configuration off that concrete type.
Add in a super-admin ower to the organization, and provider type
specific configuration that will allow access to groups information.  In
that vein, add group mappings into the organization so we can map a
user's groups to claims in the access token.