Add Groups Mapping Part 1
As it stands anyone in an organization can log in and do everything,
which we really don't want to happen, imagine if you let sales have full
admin access... We need to assign privileges based on roles assigned to
groups.  To that end, create a provider specific interface that allows
the definition of any scopes necessary to acheive the task of looking up
groups from the organization's oauth2 provider, potentially with the use
of the access token we are now in posession of during login.