IdP Group Support Part 1 (#47)
So, previously we'd extract the names, but these are unsable, so we
extract that and the dsiplay name, but then the code and authorization
headers are too big for Nginx.  So what we do now is remember the access
token and can use that for point-in-time group lookups, which is good as
it will reflect what's in the IdP.  However we may need to think about
caching for performance reasons down the line, or we could end up
blowing API limits.  Anyway, we can now extract the groups, and display
them in the UI when setting up a group.