Issuer Key Rotation (#49)
THe problem we are solving here is currently a token may be issued right
before the key pair is rotated, so be comes invalid almost immediately.
This will become more troublesome down the line with automation
requiring long lived keys (i.e. refresh tokens).  A solution is to
create keys that last for 2N, and rotate every N, so any token issued
just before a rotation will work for a time of N.  To pull this off we
need to create copies of the current keypair and the previous for
verification purposes and rotation detection.