Work Around Request Size (#80)
Login in Entra is causing the request size to be blown due to large
tokens, probably hinting they encode permissions in there rather thand
doing a DB lookup.  As a quick hack, remove the tokens from the code and
our tokens to reduce the size, as they aren't necessary.  A long term
solution will be to save these details in a DB and pass around much
smaller identifiers to allow a lookup.  We need the DB anyway for access
token single use and revocaction.