RBAC Done Properly (#94)
Where to start... what we had was an evolution.  It used interfaces in
ways they shouldn't be used, caused messy API hander code, repetition
and it didn't actually work, especially for project scoped resources.
This fixes all that by making identity the source of truth for
everything (as opposed to spreading it over core).  Rationalizes all the
functionality and just generally makes it more robust, flexible with
less hackery involved.