Make Regions Organization Scoped (#101)
After playing with the UI, it semms quite apparent that mapping a region
ID to a name, when the region is itself project scoped is quite heavy
weight and unwieldy.  To remedy this we move regions back to the
organization scope so we need only a single lookup, relying on RBAC
information to do any filtering.  This also splits up regions and
instrastructure so readers can map to regions, but not create identities
etc.