Tighten RBAC (#130)
In the old world, before service accounts, the user needed excess
premissions to achieve anything, as the token was propagated around
services.  You couldn't have a user that could just provision
Kubernetes, they needed to be able to provision identities and physical
networks too, the latter is a major concern when it comes to DoS attacks
against the platform via the API.  This also scopes endpoints so they
are easier to work with, and also adds in better RBAC for flavors and
images e.g. by going through the specific service, that can perform
domain specific filtering of usable resources.