Add Users API
This doesn't actually operate on a resource, as users themselves are
federated from another source of truth, but we can at least make a meta
API out of the data contained in groups.  This allows us to easily add
and remove users to groups, rather than having to manually do it per
group.