Add User Resource
Previously we allowed organization admins to allow anyone to join their
organization based on group membership.  This creates a user CRD which
performs ostensibly the same thing, but also allow the uesr to have
different states e.g. to be disabled or onboarded with a 2FA challenge
and response.  It can also recurd things like user activity.