Add SMTP Signup Notifications
Enables a mode of operation where new users will enter the pending state
and will be sent an email with a link to click on in order to transition
to active.  As you would expect, any emails and errors allow branding.
For the simple case it's just a case of picking up a template from a
configmap.  For redirects, it's a bit more complex as we need to
propagate client information and register some new addresses on the
client.