Tweak Identity APIs
If we have an empty list of say group IDs for a user, then the UI need
to fill this in so it can bind to it, and then yout still get a bunch of
if statements that clutter up the code.  What would be better is to
return an empty array so it's already setup.  Cleaner code at the
expense of a few extra bytes on the wire.  This should be backwards
compatible with existing code as it's just making things that are
already handled required rather than optional.