Remove Source of Privilege Escalation
It was noted that an admin user whose organization contained a
platform-administrators group could add any user or service account to
that group, thus granting privilege escalation.  This removes the idea
of a super admin group entirely, instead relying on it being defined by
the deployment and thus under review and version control.

Fixes #176.