Allow External Admin Roles (#217)
Previously only a single role could be granted to a platform admin
meaning they couldn't be granted 3rd party service permissions without
either extending the base role, or providing their own.  The former is
painful as you need to keep the base values in lockstep with the 3rd
party integration, the latter is the same problem.  This proposes
allowing the role to be defined once in the 3rd party application, and
included in the base one via a single name reference.