DIfferentiate Between Subject and Email
This is kind inconsistent in that the subject is sometimes a service
account ID and sometimes a user's email address.  If we force it to be a
user ID instead, this means lookups are quicker and less error prone,
this forces clients into using the email claim, which they should use
anyway, and may be slightly breaking.