Organization Filtering 2
Some users like to hand out platform admin like candy to a baby.  This
may result in bloated organization lists.  We already have a filtering
option so extend it so that a user from a client can pass in their
email, sourced from the ID token, to the organizations endpoint to
constrain the visible options to only those that they are actually a
member of.