Allow Service Account Self-Rotation (#276)
The core of this is to allow a service account to rotate its own key
periodically to avoid any human interaction, we do this by modifying the
RBAC to first check if the actor is a service account, and the service
account ID matches that of the requested ID.

To actually get the ID to rotate in the first place you need to get the
organization it belongs to by doing a list against the organizations
endpoint, then list the service accounts to get the ID.  This led to the
discovery that RBAC is missing on the service account list endpoint.  So
this applies the same RBAC changes to rotation, but ignores it when a
service account ID is not provided, then filters by service account ID
if it doesn't posess the full `serviceaccounts:read` privilege.  This
provides the full path to glory.

For service account tokens that have the full privilege level, they will
need to use extra logic in the client to derive the ID e.g. use a name
or tag to match, or just use the ID directly.

Fixes #275