Principal Propagation
This does a few things to enable principal propagation throughout the
system.

Defines what a principal is, organization, project and end-user identity
tuple and provides some machinery to stash and retrieve from the
context.

We make it so principals are always propagated by a system service,
otherwise they are generated for the authenticated end user of a UI/CLI
client by the authentication middleware.

We provide interfaces for clients to use.  For handlers we move the
identity bits of ObjectMeta out of core into here, then have it
implicitly fill in identity information for created Kubernetes
resources, and also provide an interface to get access to the principal
(from the context), for use with things like quota allocations.

The final interface is a reimagining of how the clients work.  At
present it's basically a copy and paste, but with generics we can
abstract away the type specific bits and always inject principal
information transparently as a HTTP header.