Principal Propagation
This does a few things to enable principal propagation throughout the
system.

Defines what a principal is, organization, project and end-user identity
tuple and provides some machinery to stash and retrieve from the
context.

Next it adds in some openapi middleware to extract the principal header
from a HTTP request, and validate it against the (validated by mTLS)
client certificate and stash it in the context.

We provide interfaces for clients to use.  For handlers we move the
identity bits of ObjectMeta out of core into here, then have it
implicitly fill in identity information for created Kubernetes
resources, and also provide an interface to get access to the principal
(from the context), for use with things like quota allocations.  The
final interface is used by controllers who consume said resources and
proxy requests on to other services, this generically extracts the
principal from a resources, encodes it and injects it into a remote HTTP
request as a header.