Allow OAuth2 Secrets from Secrets
Allows the clinet ID and secret to be read from a secret, for example
provided by external secrets operator.  Most of this is just tweaking
the interfaces to propagate the clients to where they are needed, then
just using literals or the secret.  Spotted a few library calls that can
replace some exsting configuration, and where we were specifiying things
that are implictly added.  Also fixed a typo in the API, so that needs
calling out.

Fixes #279.