Rewrite to use core/middleware.Capture
The handler responses are captured (status code, headers, body) in
order to validate them against the OpenAPI schema.

At present this is done with a wrapper for ResponseWriter; but it's
non-trivial to properly wrap ResponseWriter, as the tests in core
show. In general, wrapping http.ResponseWrite is fraught (see
https://github.com/felixge/httpsnoop/tree/v1.0.4?tab=readme-ov-file#why-this-package-exists).

This commit replaces the ad-hoc ResponseWriter wrapper, and the use of
middleware.LoggingResponseWriter which has been deprecated. Instead it
uses core/middleware.CaptureResponse to capture the response where
needed.