Test the response capture machinery
The handler responses are captured (status code, headers, body) in
order to validate them against the OpenAPI schema.

At present this is done with a wrapper for ResponseWriter; but it's
non-trivial to properly wrap ResponseWriter, as the added test shows.

To make it testable, I factored out the part where the validator runs
the next http.Handler, so it can be run in isolation.