Implement response capture with httpsnoop Wrap
Wrapping http.ResponseWrite is fraught (see
https://github.com/felixge/httpsnoop/tree/v1.0.4?tab=readme-ov-file#why-this-package-exists).

This uses httpsnoop.Wrap to wrap the response while trying to capture
the bytes written to it, etc. This overcomes some gaps in the prior
implementation -- which could surely have been papered over, but why
not use the battle-tested library to do it once and for all.

I've supplied a wrapper for ReadFrom because httpsnoop will call it if
it exists on the ResponseWriter, and the implementation in `http` does
implement it. (It was fine _not_ having it on the prior implementation,
because io.Copy would fall back to buffered writes anyway.)